requirements.txt
pillow 
tqdm
progressbar